+++ src/i18n/useT.ts
@@ -1,0 +1,7 @@
+import { useContext } from 'react';
+
+import { LangContext } from './LangContext';
+
+export function useT() {
+  return useContext(LangContext);
+}